add admin teacher authoriztion.

+++ app/Http/Controllers/Admin/TeacherController.php
@@ -12,12 +12,20 @@ use App\Models\Course;
 use App\Models\Teacher;
 use App\Http\Resources\TeacherResource;
 use App\Http\Resources\TeacherCollection;
+use Illuminate\Support\Facades\Auth;
 
 class TeacherController extends Controller
 {
 
     public function request_teachers()
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
         $teachers = ApplyTeacher::select('id', 'username', 'description')->get();
 
 
@@ -26,6 +34,13 @@ class TeacherController extends Controller
 
     public function answer_request(ApplyTeacher $request)
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
        $request->user; 
        return Inertia::render('Teachers/AnswerRequest', [
            'request' => $request
@@ -35,6 +50,13 @@ class TeacherController extends Controller
 
     public function reject_request(ApplyTeacher $request)
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
         $request->delete();
 
         return redirect()->route('teacher.requests');
@@ -42,6 +64,13 @@ class TeacherController extends Controller
 
     public function accept_request(ApplyTeacher $request)
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
         Teacher::create([
             'uuid' => Str::uuid(),
             'first_name' => $request->first_name,
@@ -64,6 +93,13 @@ class TeacherController extends Controller
 
     public function rejected_requests()
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
         $requests = ApplyTeacher::onlyTrashed()->select('id','username', 'description')->get();
         return Inertia::render('Teachers/RejectedRequests', ['requests' => $requests]);
     }
@@ -71,6 +107,13 @@ class TeacherController extends Controller
 
     public function requests_course()
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
         $requests = RequestCourse::all();
 
         return Inertia::render('Teachers/RequestsCourse', [
@@ -85,6 +128,13 @@ class TeacherController extends Controller
      */
     public function index()
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
         return Inertia::render('Teachers/Index', [
             'teachers' => new TeacherCollection(Teacher::all())
         ]);
@@ -92,6 +142,13 @@ class TeacherController extends Controller
 
     public function answer_request_course(RequestCourse $request)
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
         $request->teacher;
         $request->teacher->user;
         return Inertia::render('Teachers/AnswerRequestCourse', [ 'request' => $request ]);
@@ -99,6 +156,13 @@ class TeacherController extends Controller
 
     public function reject_request_course(RequestCourse $request)
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
         $request->delete();
 
         return redirect()->route('teacher.requests');
@@ -106,6 +170,13 @@ class TeacherController extends Controller
 
     public function accept_request_course(RequestCourse $request)
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
         Course::create([
             'sku' => Str::uuid(),
             'title' => $request->title,
@@ -127,6 +198,13 @@ class TeacherController extends Controller
 
     public function rejected_courses()
     {
+        $user = Auth::user();
+        if(!$user->can('teachers admin'))
+        {
+            abort(403);
+        }
+        
+
         $requests = RequestCourse::onlyTrashed()->get();
